Move check_yaml to the utils folder
This is to align shared code between the agent5 and agent6.
